requirements.txt
pandas
chardet
python-magic